fix: require node 14.18 or higher
Some of our upstream dependencies have started using `node:`-prefixed imports.
This is causing users who are on Node 14 _but less than the .18_ minor version to
crash due to not being able to resolve the module. While normally this would be a
breaking change, I feel it is _probably_ okay to treat is as a patch, given anyone
using Node < 14.18 will currently just have it crashing on them.